+++ src/mistralcli/__init__.py
@@ -1,9 +1,9 @@
-"""mistral4cli package."""
+"""mistralcli package."""
 
 from importlib.metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version("mistral4cli")
+    __version__ = version("mistralcli")
 except PackageNotFoundError:
     __version__ = "0.0.0"
 
@@ -12,6 +12,8 @@ from .conversation_registry import ConversationRegistry
 from .local_mistral import (
     DEFAULT_MODEL_ID,
     DEFAULT_SERVER_URL,
+    REMOTE_MEDIUM_MODEL_ID,
+    REMOTE_MODEL_ID,
     BackendKind,
     ContextConfig,
     ConversationConfig,
@@ -21,6 +23,7 @@ from .local_mistral import (
     build_client,
     get_health,
     list_models,
+    normalize_remote_model_id,
     remote_api_key_available,
 )
 from .session import MistralCodingSession, MistralSession, PendingConversationSettings
@@ -28,6 +31,8 @@ from .session import MistralCodingSession, MistralSession, PendingConversationSe
 __all__ = [
     "DEFAULT_MODEL_ID",
     "DEFAULT_SERVER_URL",
+    "REMOTE_MEDIUM_MODEL_ID",
+    "REMOTE_MODEL_ID",
     "BackendKind",
     "ContextConfig",
     "ConversationConfig",
@@ -43,5 +48,6 @@ __all__ = [
     "get_health",
     "list_models",
     "main",
+    "normalize_remote_model_id",
     "remote_api_key_available",
 ]